requirements.txt
torch
torchvision
Pillow
grpcio 
grpcio-tools